+++ CHANGELOG.md
@@ -27,7 +27,7 @@
 ## v0.0.7
 - fixed /app/tasmoadmin/.docker to disable SelfUpdate
 - restyle show more
-
+- add send command within device list #100
 
 ## v0.0.6   
 ### 2018-05-23